// src/core/enums/cu-job.enum.ts
export enum CuJobEnum {
  Facilitator = 'Facilitator',
  ResearchExpert = 'Research Expert',
  ProjectLead = 'Project Lead',
  DataExpert = 'Data Expert',
  TechExpert = 'Tech expert',
  TeamLead = 'Team Lead',
  Contributor = 'Contributor',
  Legal = 'Legal',
  Product = 'Product',
  Developer = 'Developer',
  DataAnalyst = 'Data Analyst',
}